Fix flaky unit tests (#290)
