+++ src/test/app/forms/models/dateOfBirth.ts
@@ -83,7 +83,7 @@ describe('DateOfBirth', () => {
       it('should reject date of birth with age over 150', () => {
         const over150YearsAgo = moment().subtract(151, 'years')
 
-        const errors = validator.validateSync(dateOfBirth(over150YearsAgo.year(), over150YearsAgo.month(), over150YearsAgo.date()))
+        const errors = validator.validateSync(dateOfBirth(over150YearsAgo.year(), over150YearsAgo.month() + 1, over150YearsAgo.date()))
 
         expect(errors.length).to.equal(1)
         expectValidationError(errors, ValidationErrors.DATE_NOT_VALID)
